Time: 12 ms (97.65%) | Memory: 47.1 MB (14.42%) - LeetSync

+++ 561-array-partition/array-partition.java
@@ -1,10 +1,10 @@
 class Solution {
     public int arrayPairSum(int[] nums) {
         Arrays.sort(nums);
-        int result = 0;
+        long result = 0;
         for(int i = 0; i<nums.length; i = i+2){
             result = result+nums[i];
         }
-        return result;
+        return (int) result;
     }
 }
